Move schemas.py to root directory
